requirements_2.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
Django==4.0.2
django-environ==0.8.1      
djangorestframework==3.13.1
executing==0.8.3
ipython==8.1.1
jedi==0.18.1
Jinja2==3.0.3
jsonpickle==2.1.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.3   
networkx==2.7.1
numpy==1.22.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.28
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
pyvis==0.1.9
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
tzdata==2021.5
wcwidth==0.2.5